requirements.txt
click==8.1.3
colorama==0.4.5
et-xmlfile==1.1.0
Flask==2.2.2
Flask-WTF==1.0.1
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.2
openpyxl==3.0.10
pandas==1.4.3
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.2.1
six==1.16.0
Werkzeug==2.2.2
WTForms==3.0.1
xlrd==2.0.1
